fixed it so the jar file correctly uses the resource file for the floorPlan as well as refactored the logger to be inside of the com.cleansweep package
