+++ src/main/java/com/cleansweep/ControlSystem/Vacuum.java
@@ -1,7 +1,7 @@
 package com.cleansweep.ControlSystem;
 
 
-import main.java.Logger;
+import com.cleansweep.Logger;
 
 import java.lang.reflect.Array;
