+++ src/main/resources/floorPlan.json
